refactor(`JekyllSiteProvider`, `PageItem`): seperate `PageItem` from `JekyllSiteProvider`
